Django_Rest_Oauth2_Backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.4.0
cryptography==41.0.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==5.0.1
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.3
django-logging-json==1.15
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.3
drf-spectacular==0.27.0
drf-standardized-errors==0.12.5
ecdsa==0.18.0
elastic-transport==8.12.0
elasticsearch==8.12.0
exceptiongroup==1.2.0
flake8==6.1.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-mock==3.12.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
wrapt==1.16.0
